// ecommerce-platform-frontend/src/app/dto/email/email-objects-enum.ts
export enum EmailObjectsEnum {
  REGISTRATION = 'REGISTRATION',
  EMAIL_PASSWORD_CHANGE_CONFIRMATION = 'EMAIL_PASSWORD_CHANGE_CONFIRMATION',
  PASSWORD_RESET = 'PASSWORD_RESET',
  ORDER_CONFIRMATION = 'ORDER_CONFIRMATION',
  NEWSLETTER = 'NEWSLETTER',
  CUSTOM = 'CUSTOM',
  ORDER_STATUS_UPDATE = 'ORDER_STATUS_UPDATE'
}